Sqlmodel update (#18)

+++ project-api/app/crud/v1/base.py
@@ -1,43 +1,35 @@
-from typing import Any, Dict, Generic, List, Optional, Type, TypeVar, Union
+from typing import Any, Dict, Generic, Optional, Type, TypeVar, Union
 
 from fastapi.encoders import jsonable_encoder
-from pydantic import BaseModel
-from sqlalchemy.orm import Session
-from sqlalchemy import select, func
 from fastapi_pagination.ext.sqlalchemy import paginate
 from fastapi_pagination import Page
+from sqlmodel import SQLModel, select, Session, func
 
-from app.db.base_class import Base
-
-ModelType = TypeVar("ModelType", bound=Base)
-CreateSchemaType = TypeVar("CreateSchemaType", bound=BaseModel)
-UpdateSchemaType = TypeVar("UpdateSchemaType", bound=BaseModel)
+ModelType = TypeVar("ModelType", bound=SQLModel)
+CreateSchemaType = TypeVar("CreateSchemaType", bound=SQLModel)
+UpdateSchemaType = TypeVar("UpdateSchemaType", bound=SQLModel)
 
 
 class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
     def __init__(self, model: Type[ModelType]):
         """
         CRUD object with default methods to Create, Read, Update, Delete (CRUD).
-
-        **Parameters**
-
-        * `model`: A SQLAlchemy model class
-        * `schema`: A Pydantic model (schema) class
         """
         self.model = model
 
     def get(self, db: Session, id: Any) -> Optional[ModelType]:
-        return db.scalars(select(self.model).where(self.model.id == id)).first()
+        return db.get(self.model, id)
 
     def count(self, db: Session) -> int:
-        return db.execute(select(func.count()).select_from(self.model)).scalar()
+        result = db.exec(select(func.count()).select_from(self.model)).first()
+        count = result[0] if result else 0
+        return count
 
     def get_multi(self, db: Session) -> Page[ModelType]:
         return paginate(db, select(self.model).order_by(self.model.id.desc()))
 
     def create(self, db: Session, *, obj_in: CreateSchemaType) -> ModelType:
-        obj_in_data = jsonable_encoder(obj_in)
-        db_obj = self.model(**obj_in_data)  # type: ignore
+        db_obj = self.model(**obj_in.model_dump())
         db.add(db_obj)
         db.commit()
         db.refresh(db_obj)
@@ -50,21 +42,21 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         db_obj: ModelType,
         obj_in: Union[UpdateSchemaType, Dict[str, Any]],
     ) -> ModelType:
-        obj_data = jsonable_encoder(db_obj)
-        if isinstance(obj_in, dict):
-            update_data = obj_in
-        else:
-            update_data = obj_in.dict(exclude_unset=True)
-        for field in obj_data:
-            if field in update_data:
-                setattr(db_obj, field, update_data[field])
+        update_data = (
+            obj_in.model_dump(exclude_unset=True)
+            if isinstance(obj_in, SQLModel)
+            else obj_in
+        )
+        for field, value in update_data.items():
+            setattr(db_obj, field, value)
         db.add(db_obj)
         db.commit()
         db.refresh(db_obj)
         return db_obj
 
-    def remove(self, db: Session, *, id: int) -> ModelType:
-        obj = db.query(self.model).get(id)
-        db.delete(obj)
-        db.commit()
+    def remove(self, db: Session, *, id: int) -> Optional[ModelType]:
+        obj = db.get(self.model, id)
+        if obj:
+            db.delete(obj)
+            db.commit()
         return obj
